requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==3.2
django-allauth==0.54.0
django-celery-results==2.5.1
django-filter==23.2
django-redis==5.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.2
documentation==0.0.1
drf-yasg==1.21.7
drf-yasg2==1.19.4
Flask==2.3.2
Flask-Cors==4.0.0
flask-restplus==0.13.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kombu==5.3.1
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
swagger-ui==0.1.2
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
